requirements.txt
groq==0.*
pytest
python-dotenv
